Remove weird circular dependency between Serializable and S.ComponentMap
