+++ Farquaad/include/Farquaad/Core.hpp
@@ -3,4 +3,5 @@
 #pragma once
 
 #include <Farquaad/Core/Serializable.hpp>
+#include <Farquaad/Core/SeralizeableComponentMap.h>
 #include <Farquaad/Core/ComponentSerializer.h>
